@@ -18,33 +18,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.1+4
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_map: ^0.9.0
+  flutter_map: ^0.12.0
   latlong: ^0.6.1
-  location: ^3.0.2
-  provider: ^4.1.3
-  dio: ^3.0.9
-  property_change_notifier: ^0.2.0
-  rxdart: ^0.24.1
-  dio_http_cache: ^0.2.7
+  location: ^4.1.1
+  provider: ^5.0.0
+  dio: ^4.0.0
+  rxdart: ^0.26.0
+  # dio_http_cache: ^0.2.7
   sliding_up_panel: ^1.0.2
-  flutter_icons: ^1.1.0
-  flutter_compass: ^0.4.2+1
-  url_launcher: ^5.5.0
+  flutter_feather_icons: ^2.0.0
+  flutter_compass: ^0.6.0
+  url_launcher: ^6.0.3
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  in_app_purchase: ^0.3.4+5
+  cupertino_icons: ^1.0.2
+  in_app_purchase: ^0.5.2
+
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: "^0.7.3"
+  flutter_launcher_icons: ^0.8.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
