@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 class Credits extends StatelessWidget {
-  const Credits({Key key}) : super(key: key);
+  const Credits({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
