@@ -2,15 +2,15 @@ import 'package:flutter/foundation.dart';
 import 'dart:async';
 
 class Debouncer {
-  final int milliseconds;
-  VoidCallback action;
-  Timer _timer;
+  late final int milliseconds;
+  VoidCallback? action;
+  Timer? _timer;
 
-  Debouncer({this.milliseconds});
+  Debouncer({required this.milliseconds});
 
   run(VoidCallback action) {
     if (_timer != null) {
-      _timer.cancel();
+      _timer!.cancel();
     }
 
     _timer = Timer(Duration(milliseconds: milliseconds), action);
